Create `config.ui` subpackage
